Seperated __init__ to seperate modules. __init__ still exportst Graph like the documentation states
